+++ src/main/java/ca/uhnresearch/pughlab/tracker/audit/AuditLogEventHandler.java
@@ -1,10 +1,9 @@
 package ca.uhnresearch.pughlab.tracker.audit;
 
-import com.fasterxml.jackson.databind.JsonNode;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 
-import ca.uhnresearch.pughlab.tracker.dao.AuditLogRepository;
 import ca.uhnresearch.pughlab.tracker.dto.AuditLogRecord;
+import ca.uhnresearch.pughlab.tracker.dao.AuditLogRepository;
 import ca.uhnresearch.pughlab.tracker.events.Event;
 import ca.uhnresearch.pughlab.tracker.events.EventHandler;
 import ca.uhnresearch.pughlab.tracker.events.RedactedJsonNode;
@@ -13,19 +12,19 @@ public class AuditLogEventHandler implements EventHandler {
 	
 	private AuditLogRepository repository;
 
+	/**
+	 * Handles a message by writing a new audit log record.
+	 * @param event the event
+	 * @param scope the event scope
+	 */
 	@Override
 	public void sendMessage(Event event, String scope) {
 		
-		final ObjectNode parameters = event.getData().getParameters().deepCopy();
-		final JsonNode oldValue = parameters.get("old");
-		if (oldValue != null && oldValue instanceof RedactedJsonNode) {
-			parameters.replace("old", ((RedactedJsonNode) oldValue).getValue());
-		}
-		final JsonNode newValue = parameters.get("new");
-		if (newValue != null && newValue instanceof RedactedJsonNode) {
-			parameters.replace("new", ((RedactedJsonNode) newValue).getValue());
-		}
-
+		// Strip redaction here, so we write clear. There's no need to redact when writing to the
+		// audit log, although it's handy for scripting and event handling in other forms, like
+		// socket data.
+		final ObjectNode parameters = RedactedJsonNode.redactedToClear(event.getData().getParameters());
+		
 		AuditLogRecord record = new AuditLogRecord();
 		if (parameters.has("study_id")) {
 			record.setStudyId(parameters.get("study_id").asInt());
@@ -43,6 +42,7 @@ public class AuditLogEventHandler implements EventHandler {
 	}
 
 	/**
+	 * Retrieves the audit log repository.
 	 * @return the repository
 	 */
 	public AuditLogRepository getRepository() {
@@ -50,6 +50,7 @@ public class AuditLogEventHandler implements EventHandler {
 	}
 
 	/**
+	 * Sets the audit log repository.
 	 * @param repository the repository to set
 	 */
 	public void setRepository(AuditLogRepository repository) {