+++ src/main/java/ca/uhnresearch/pughlab/tracker/dao/StudyRepository.java
@@ -20,7 +20,7 @@ public interface StudyRepository {
 	List<Study> getAllStudies();
 	
 	/**
-	 * Retrieves a single specified study from the repository
+	 * Retrieves a single specified study from the repository.
 	 * @return a study
 	 */
 	Study getStudy(String name);
@@ -108,9 +108,14 @@ public interface StudyRepository {
 	List<ObjectNode> getCaseData(StudyCaseQuery query, List<? extends Attributes> attributes);
 	
 	/**
+	 * Deletes a set of cases from the repository.
+	 */
+	void deleteCases(StudyCaseQuery query, String userName) throws RepositoryException;
+
+	/**
 	 * Builds a new study case query, which can be transformed into a set of cases.
 	 * This is always initialized to a single study. 
-	 * @return
+	 * @return a new StudyCaseQuery
 	 */
 	StudyCaseQuery newStudyCaseQuery(Study study);
 
@@ -118,7 +123,7 @@ public interface StudyRepository {
 	 * Adds a view to the case matcher, which might filter out a bunch of data that should not
 	 * be visible. These will typically be row filters, as cases are being selected by the
 	 * case query, not attributes. 
-	 * @return
+	 * @return a modified StudyCaseQuery
 	 */
 	StudyCaseQuery addViewCaseMatcher(StudyCaseQuery query, View view);
 
@@ -128,7 +133,7 @@ public interface StudyRepository {
 	 * and can be used later.
 	 * @param attribute
 	 * @param value
-	 * @return
+	 * @return a modified StudyCaseQuery
 	 */
 	StudyCaseQuery addStudyCaseMatcher(StudyCaseQuery query, String attribute, String value);
 	
@@ -136,9 +141,9 @@ public interface StudyRepository {
 	 * Adds a new matching element to a study case filter. Only string values 
 	 * are typically allowed here, currently. A new study case query is returned.
 	 * and can be used later.
-	 * @param attribute
-	 * @param value
-	 * @return
+	 * @param query
+	 * @param caseId
+	 * @return a modified StudyCaseQuery
 	 */
 	StudyCaseQuery addStudyCaseSelector(StudyCaseQuery query, Integer caseId);
 
@@ -146,9 +151,9 @@ public interface StudyRepository {
 	 * Adds a new matching element to a study case filter. Only string values 
 	 * are typically allowed here, currently. A new study case query is returned
 	 * and can be used later.
-	 * @param attribute
-	 * @param value
-	 * @return
+	 * @param query
+	 * @param filter
+	 * @return a modified StudyCaseQuery
 	 */
 	StudyCaseQuery addStudyCaseFilterSelector(StudyCaseQuery query, ObjectNode filter);
 
@@ -157,7 +162,7 @@ public interface StudyRepository {
 	 * process.
 	 * @param query
 	 * @param attribute
-	 * @return
+	 * @return a modified StudyCaseQuery
 	 */
 	StudyCaseQuery subcases(StudyCaseQuery query, String attribute);
 	
@@ -165,7 +170,7 @@ public interface StudyRepository {
 	 * Applies a CasePager to a query, returning a new query.
 	 * @param query
 	 * @param pager
-	 * @return
+	 * @return a modified StudyCaseQuery
 	 */
 	StudyCaseQuery applyPager(StudyCaseQuery query, CasePager pager);
 
@@ -173,7 +178,7 @@ public interface StudyRepository {
 	 * Retrieves a single specified case for a study and view from the repository.
 	 * @return a case
 	 */
-	Cases getStudyCase(Study study, View view, Integer caseId);
+	Cases getStudyCase(Study study, Integer caseId);
 	
 	/**
 	 * /**
@@ -183,11 +188,11 @@ public interface StudyRepository {
 	 * triggered by other changes, and generate notifications. 
 	 * 
 	 * @param study
-	 * @param view
 	 * @param cases
+	 * @param userName
 	 * @param state
 	 */
-	void setStudyCaseState(Study study, View view, Cases cases, String userName, String state);
+	void setStudyCaseState(Study study, Cases cases, String userName, String state);
 	
 	/**
 	 * Makes a new, empty, case. The new case will be added after afterCase
@@ -196,18 +201,17 @@ public interface StudyRepository {
 	 * @return the case identifier
 	 * @throws RepositoryException
 	 */
-	Cases newStudyCase(Study study, View view, String userName, Cases afterCase) throws RepositoryException;
+	Cases newStudyCase(Study study, String userName, Cases afterCase) throws RepositoryException;
 	
 	/**
 	 * Makes a new, empty, case. The new case will be added at the
 	 * beginning of the case list, which is only sometimes what you want.
 	 * @param study
-	 * @param view
 	 * @param userName
-	 * @return
-	 * @throws RepositoryException
+	 * @return a new case
+	 * @throws RepositoryException 
 	 */
-	Cases newStudyCase(Study study, View view, String userName) throws RepositoryException;
+	Cases newStudyCase(Study study, String userName) throws RepositoryException;
 	
 	/**
 	 * Selects which attributes will be returned in a query.