+++ src/main/java/ca/uhnresearch/pughlab/tracker/dto/ViewPermissions.java
@@ -2,56 +2,62 @@ package ca.uhnresearch.pughlab.tracker.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
-public class ViewPermissions extends AdministratorPermissions {
+public class ViewPermissions extends StudyPermissions {
 
 	public ViewPermissions() { }
 	
 	/**
-	 * @return the readAllowed
+	 * Retrieves the read permission.
+	 * @return the read
 	 */
 	@JsonProperty
-	public Boolean getReadAllowed() {
-		return readAllowed;
+	public Boolean getRead() {
+		return read;
 	}
 
 	/**
-	 * @param readAllowed the readAllowed to set
+	 * Sets the read permission.
+	 * @param read the read to set
 	 */
-	public void setReadAllowed(Boolean readAllowed) {
-		this.readAllowed = readAllowed;
+	public void setRead(Boolean read) {
+		this.read = read;
 	}
 
 	/**
-	 * @return the writeAllowed
-	 */
-	@JsonProperty
-	public Boolean getWriteAllowed() {
-		return writeAllowed;
-	}
-
-	/**
-	 * @param writeAllowed the writeAllowed to set
-	 */
-	public void setWriteAllowed(Boolean writeAllowed) {
-		this.writeAllowed = writeAllowed;
-	}
-
-	/**
-	 * @return the downloadAllowed
+	 * Retrieves the write permission.
+	 * @return the write
 	 */
 	@JsonProperty
-	public Boolean getDownloadAllowed() {
-		return downloadAllowed;
+	public Boolean getWrite() {
+		return write;
 	}
 
 	/**
-	 * @param downloadAllowed the downloadAllowed to set
+	 * Sets the write permission.
+	 * @param write the write to set
 	 */
-	public void setDownloadAllowed(Boolean downloadAllowed) {
-		this.downloadAllowed = downloadAllowed;
+	public void setWrite(Boolean write) {
+		this.write = write;
 	}
 
-	private Boolean readAllowed = false;
-	private Boolean writeAllowed = false;
-	private Boolean downloadAllowed = false;
+	/**
+	 * Retrieves the download permission.
+	 * @return the download
+	 */
+	@JsonProperty
+	public Boolean getDownload() {
+		return download;
+	}
+
+	/**
+	 * Sets the download permission.
+	 * @param download the download to set
+	 */
+	public void setDownload(Boolean download) {
+		this.download = download;
+	}
+
+	private Boolean read = false;
+	private Boolean write = false;
+	private Boolean download = false;
 }