+++ src/main/java/ca/uhnresearch/pughlab/tracker/resource/EntityResource.java
@@ -2,9 +2,14 @@ package ca.uhnresearch.pughlab.tracker.resource;
 
 import java.util.List;
 
+import org.apache.shiro.SecurityUtils;
+import org.apache.shiro.subject.PrincipalCollection;
+import org.apache.shiro.subject.Subject;
+import org.restlet.Request;
 import org.restlet.data.Status;
 import org.restlet.ext.jackson.JacksonRepresentation;
 import org.restlet.representation.Representation;
+import org.restlet.resource.Delete;
 import org.restlet.resource.Get;
 import org.restlet.resource.ResourceException;
 import org.slf4j.Logger;
@@ -12,6 +17,8 @@ import org.slf4j.LoggerFactory;
 
 import com.fasterxml.jackson.databind.node.ObjectNode;
 
+import ca.uhnresearch.pughlab.tracker.dao.NotFoundException;
+import ca.uhnresearch.pughlab.tracker.dao.RepositoryException;
 import ca.uhnresearch.pughlab.tracker.dao.StudyCaseQuery;
 import ca.uhnresearch.pughlab.tracker.dto.EntityResponse;
 import ca.uhnresearch.pughlab.tracker.dto.Study;
@@ -19,7 +26,10 @@ import ca.uhnresearch.pughlab.tracker.dto.View;
 
 public class EntityResource extends StudyRepositoryResource<EntityResponse> {
 	
-	private final Logger logger = LoggerFactory.getLogger(getClass());
+	/**
+	 * A logger
+	 */
+	private final Logger logger = LoggerFactory.getLogger(EntityResource.class);
 	
     @Get("json")
     public Representation getResource() {
@@ -28,6 +38,31 @@ public class EntityResource extends StudyRepositoryResource<EntityResponse> {
         return new JacksonRepresentation<EntityResponse>(response);
     }
 
+	@Delete()
+	public void deleteResource()  {
+		Subject currentUser = SecurityUtils.getSubject();
+		
+		Request request = getRequest();
+    	Study study = RequestAttributes.getRequestStudy(request);
+    	StudyCaseQuery query = RequestAttributes.getRequestCaseQuery(request);
+
+    	Boolean deletePermitted = currentUser.isPermitted(study.getName() + ":delete");
+    	if (! deletePermitted) {
+    		throw new ResourceException(Status.CLIENT_ERROR_FORBIDDEN);
+    	}
+
+		PrincipalCollection principals = currentUser.getPrincipals();
+		String user = principals.getPrimaryPrincipal().toString();
+
+    	try {
+			getRepository().deleteCases(query, user);
+    	} catch (NotFoundException e) {
+			throw new ResourceException(Status.CLIENT_ERROR_NOT_FOUND);
+		} catch (RepositoryException e) {
+			throw new ResourceException(Status.CLIENT_ERROR_BAD_REQUEST);
+		};
+	}
+	
 	@Override
 	public void buildResponseDTO(EntityResponse dto) {
 		super.buildResponseDTO(dto);