+++ src/test/java/ca/uhnresearch/pughlab/tracker/dao/impl/StudyRepositoryImplTest.java
@@ -8,13 +8,12 @@ import java.util.ArrayList;
 import java.util.List;
 
 import static org.easymock.EasyMock.*;
-import org.hamcrest.Matchers;
 
+import org.hamcrest.Matchers;
 import org.apache.commons.lang3.builder.EqualsBuilder;
 import org.junit.Rule;
 import org.junit.Test;
 import org.junit.Assert;
-
 import org.junit.rules.ExpectedException;
 import org.junit.runner.RunWith;
 import org.slf4j.Logger;
@@ -425,8 +424,7 @@ public class StudyRepositoryImplTest {
 	@Rollback(true)
 	public void testSingleCase() {
 		Study study = studyRepository.getStudy("DEMO");
-		View view = studyRepository.getStudyView(study, "track");
-		Cases caseValue = studyRepository.getStudyCase(study, view, 1);
+		Cases caseValue = studyRepository.getStudyCase(study, 1);
 		Assert.assertNotNull(caseValue);
 		Assert.assertEquals(1, caseValue.getId().intValue());
 	}
@@ -436,8 +434,7 @@ public class StudyRepositoryImplTest {
 	@Rollback(true)
 	public void testSingleMissingCase() {
 		Study study = studyRepository.getStudy("DEMO");
-		View view = studyRepository.getStudyView(study, "track");
-		Cases caseValue = studyRepository.getStudyCase(study, view, 100);
+		Cases caseValue = studyRepository.getStudyCase(study, 100);
 		Assert.assertNull(caseValue);
 	}
 
@@ -446,8 +443,7 @@ public class StudyRepositoryImplTest {
 	@Rollback(true)
 	public void testSingleFromDifferentStudy() {
 		Study study = studyRepository.getStudy("DEMO");
-		View view = studyRepository.getStudyView(study, "track");
-		Cases caseValue = studyRepository.getStudyCase(study, view, 22);
+		Cases caseValue = studyRepository.getStudyCase(study, 22);
 		Assert.assertNull(caseValue);
 	}
 
@@ -1761,16 +1757,15 @@ public class StudyRepositoryImplTest {
 	@Rollback(true)
 	public void testNewCase() throws RepositoryException {
 		Study study = studyRepository.getStudy("DEMO");
-		View view = studyRepository.getStudyView(study, "track");
 
-		Cases newCase = studyRepository.newStudyCase(study, view, "test");
+		Cases newCase = studyRepository.newStudyCase(study, "test");
 		Assert.assertNotNull(newCase);
 		Assert.assertNotNull(newCase.getId());
 		Assert.assertNotNull(newCase.getStudyId());
 		
 		// And now let's dig out the new case -- mainly to check that we can actually
 		// follow this identifier.
-		Cases caseValue = studyRepository.getStudyCase(study, view, newCase.getId());
+		Cases caseValue = studyRepository.getStudyCase(study, newCase.getId());
 		Assert.assertNotNull(caseValue);
 		Assert.assertEquals(newCase.getId(), caseValue.getId());
 	}
@@ -1783,25 +1778,24 @@ public class StudyRepositoryImplTest {
 	@Rollback(true)
 	public void testNewCaseOrdering() throws RepositoryException {
 		Study study = studyRepository.getStudy("DEMO");
-		View view = studyRepository.getStudyView(study, "track");
 		
-		Cases foundCase = studyRepository.getStudyCase(study, view, 10);
+		Cases foundCase = studyRepository.getStudyCase(study, 10);
 		Integer foundCaseOrder = foundCase.getOrder();
 
-		Cases newCase = studyRepository.newStudyCase(study, view, "test", foundCase);
+		Cases newCase = studyRepository.newStudyCase(study, "test", foundCase);
 		Assert.assertNotNull(newCase);
 		Assert.assertNotNull(newCase.getId());
 		Assert.assertNotNull(newCase.getStudyId());
 		
 		// And now let's dig out the new case -- mainly to check that we can actually
 		// follow this identifier.
-		Cases caseValue = studyRepository.getStudyCase(study, view, newCase.getId());
+		Cases caseValue = studyRepository.getStudyCase(study, newCase.getId());
 		Assert.assertNotNull(caseValue);
 		Assert.assertEquals(newCase.getId(), caseValue.getId());
 		Assert.assertEquals(foundCaseOrder, caseValue.getOrder());
 		
 		// And check we've bumped the order
-		Cases refoundCase = studyRepository.getStudyCase(study, view, foundCase.getId());
+		Cases refoundCase = studyRepository.getStudyCase(study, foundCase.getId());
 		Assert.assertThat(caseValue.getOrder(), Matchers.lessThan(refoundCase.getOrder()));
 		Assert.assertThat(foundCase.getOrder(), Matchers.not(refoundCase.getOrder()));
 	}
@@ -1809,19 +1803,17 @@ public class StudyRepositoryImplTest {
 	/**
 	 * Simple test of adding a number of attributes as well as deleting.
 	 */
-	@SuppressWarnings("unchecked")
 	@Test
 	@Transactional
 	@Rollback(true)
 	public void testFailingNewCase() throws RepositoryException {
 		Study study = studyRepository.getStudy("DEMO");
-		View view = studyRepository.getStudyView(study, "track");
 		
 		QueryDslJdbcTemplate mockTemplate = createMock(QueryDslJdbcTemplate.class);
 		expect(mockTemplate.newSqlQuery()).andStubReturn(studyRepository.getTemplate().newSqlQuery());
-		expect(mockTemplate.queryForObject(anyObject(SQLQuery.class), anyObject(Expression.class))).andStubReturn(null);
+		expect(mockTemplate.queryForObject(anyObject(SQLQuery.class), (Expression<?>) anyObject(Expression.class))).andStubReturn(null);
 		expect(mockTemplate.update(eq(QCases.cases), anyObject(SqlUpdateCallback.class))).andStubReturn(new Long(1));
-		expect(mockTemplate.insertWithKey(anyObject(RelationalPath.class), anyObject(SqlInsertWithKeyCallback.class))).andStubReturn(null);
+		expect(mockTemplate.insertWithKey((RelationalPath<?>) anyObject(RelationalPath.class), (SqlInsertWithKeyCallback<?>) anyObject(SqlInsertWithKeyCallback.class))).andStubReturn(null);
 		replay(mockTemplate);
 		
 		thrown.expect(InvalidValueException.class);
@@ -1831,7 +1823,7 @@ public class StudyRepositoryImplTest {
 		studyRepository.setTemplate(mockTemplate);
 
 		try {
-			studyRepository.newStudyCase(study, view, "test");
+			studyRepository.newStudyCase(study, "test");
 		} finally {
 			studyRepository.setTemplate(originalTemplate);
 		}
@@ -1857,10 +1849,9 @@ public class StudyRepositoryImplTest {
 	@Rollback(true)
 	public void testNewCaseWithoutManager() throws RepositoryException {
 		Study study = studyRepository.getStudy("DEMO");
-		View view = studyRepository.getStudyView(study, "track");
-		Cases caseValue = studyRepository.getStudyCase(study, view, 7);
+		Cases caseValue = studyRepository.getStudyCase(study, 7);
 		
-		studyRepository.setStudyCaseState(study, view, caseValue, "morag", "pending");
+		studyRepository.setStudyCaseState(study, caseValue, "morag", "pending");
 		
 		// Check we now have an audit log entry
 		CasePager pager = new CasePager();
@@ -1876,7 +1867,7 @@ public class StudyRepositoryImplTest {
 		Assert.assertEquals("pending", entry.get("eventArgs").get("state").asText());
 		
 		// Check a re-read gets the new state
-		Cases foundValue = studyRepository.getStudyCase(study, view, 15);
+		Cases foundValue = studyRepository.getStudyCase(study, 15);
 		Assert.assertEquals("pending", foundValue.getState());
 	}
 	
@@ -2255,5 +2246,52 @@ public class StudyRepositoryImplTest {
 		List<ObjectNode> dataList = studyRepository.getCaseData(filteredQuery, view);
 		Assert.assertNotNull(dataList);
 		Assert.assertEquals(16, dataList.size());
+	}
+	
+	/**
+	 * Tests that cases can be deleted using the studyRepository
+	 * @throws RepositoryException
+	 */
+	@Test
+	@Transactional
+	@Rollback(true)
+	public void testDeleteCase() throws RepositoryException {
+		Study study = studyRepository.getStudy("DEMO");
+		View view = studyRepository.getStudyView(study, "track");
+		
+		// First check the data exists
+		StudyCaseQuery query = studyRepository.newStudyCaseQuery(study);
+		query = studyRepository.addStudyCaseSelector(query, 1);
+		
+		List<ObjectNode> data = studyRepository.getCaseData(query, view);
+		Assert.assertNotNull(data);
+		Assert.assertEquals(1,  data.size());
+		
+		query = studyRepository.newStudyCaseQuery(study);
+		query = studyRepository.addStudyCaseSelector(query, 1);
+
+		// Try the deletion
+		studyRepository.deleteCases(query, "morag");
+		
+		// Now generate the query again, and confirm we can't find it
+		query = studyRepository.newStudyCaseQuery(study);
+		query = studyRepository.addStudyCaseSelector(query, 1);
+		
+		data = studyRepository.getCaseData(query, view);
+		
+		Assert.assertNotNull(data);
+		Assert.assertEquals(0,  data.size());		
+		
+		CasePager pager = new CasePager();
+		pager.setOffset(0);
+		pager.setLimit(5);
+		List<JsonNode> auditEntries = auditLogRepository.getAuditData(study, pager);
+		
+		Assert.assertEquals(1, auditEntries.size());
+		JsonNode entry = auditEntries.get(0);
+		Assert.assertEquals("delete", entry.get("eventType").asText());
+		JsonNode entryData = entry.get("eventArgs").get("data");
+		
+		Assert.assertEquals("DEMO-01", entryData.get("patientId").asText());
 	}
 }